Find ROIs based on cell type-specific densities via graph-based method.

Find ROIs based on cell type-specific densities via graph-based method.

Usage

findROI(
  spe,
  coi,
  probs = 0.85,
  ngrid.min = 20,
  method = "walktrap",
  diag.nodes = FALSE,
  sequential.roi.name = TRUE,
  ...
)

Arguments

spe

A SpatialExperiment object.

coi

A character vector of cell types of interest (COIs).

probs

A numeric scalar. The threshold of proportion that used to filter grid by density. Default to 0.85.

ngrid.min

An integer. The minimum number of grids required for defining a ROI. Default to 20.

method

The community dectection method to be used, either walktrap or connected. Default to walktrap.

diag.nodes

Logical. Set this to TRUE to allow diagonal grid points to be adjacent nodes.

sequential.roi.name

Logical. Set this to FALSE if you want the original ROI name before filtering are retained.

...

Other parameters that passed to walktrap.community.

Value

A SpatialExperiment object.

Examples

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")