Spatial cell-type inter-correlation by density in R.

scider implements functions to analyse spatial transcriptomics data with cell type annotations by performing cell type correlation via density estimation and cell type co-localization via real number distance. Functions include density estimation, statistical modelling and visualizations.

Details

scider uses SpatialExperiment objects as the main infrastructure, which can easily be integrated with a wide variety of Bioconductor packages.

Author

Ning Liu liu.n@wehi.edu.au, Mengbo Li li.me@wehi.edu.au, Yunshun Chen yuchen@wehi.edu.au